refactor: rename to .ts & refactor indents in appView file
